Day 1 Part 1
This solves the solution to Day 1 part 1, which requires finding two
numbers in a list of entries that sum to 2020, and multiplying them
together.

The solution reads in the included data, and then sorts the list of
numbers. Then a copy is made so that we can easily traverse the data
from both the min and the max point.